+++ lib/src/home/view/screens/home_page_small.dart
@@ -6,11 +6,17 @@ import 'package:andrious/main.dart';
 import 'package:andrious/src/view.dart';
 
 class HomePageSmall extends WebPage<HomePageSmall> {
-  HomePageSmall({Key? key})
+  factory HomePageSmall({Key? key}) => _this ??= HomePageSmall._(key);
+  HomePageSmall._(Key? key)
       : super(
-            controller: HomePageSmallController<HomePageSmall>(
-                physics: const AlwaysScrollableScrollPhysics()),
-            key: key);
+          key: key,
+          controller: HomePageSmallController<HomePageSmall>(
+              physics: const AlwaysScrollableScrollPhysics()),
+          coverBanner: false,
+          accessBar: false,
+          bottomBar: true,
+        );
+  static HomePageSmall? _this;
 
   @override
   String get title => 'Andrious Solutions Ltd.';
@@ -19,6 +25,93 @@ class HomePageSmall extends WebPage<HomePageSmall> {
   List<Widget> children04(BuildContext context) {
     final con = controller as HomePageSmallController;
     return con.children04(context);
+  }
+
+  @override
+  StackWidgetProperties? screenOverlay(BuildContext context) {
+    //
+    final con = controller as HomePageSmallController;
+
+    // Don't cotinue if the overlay is not to be shown.
+    if (!con.showOverlay) {
+      return null;
+    }
+
+    final _screenSize = MyApp.screenSize;
+    final _smallScreen = MyApp.inSmallScreen;
+
+    final String msg = _smallScreen
+        ? 'This website uses cookies for your best experience on my website'
+        : 'This website uses cookies to ensure you get the best experience on my website.';
+
+    final widget = Padding(
+      padding: const EdgeInsets.all(10),
+      child: Container(
+        //           height: _scrollerHeight,
+        width: _screenSize.width,
+        margin: const EdgeInsets.only(
+          left: 4,
+          right: 4,
+          top: 8,
+          bottom: 8,
+        ),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          border: Border.all(width: 2),
+          borderRadius: const BorderRadius.all(
+            Radius.circular(3),
+          ),
+        ),
+        child: SizedBox(
+          width: _screenSize.width * 0.6,
+          height: _screenSize.height * 0.1,
+          child: Row(mainAxisSize: MainAxisSize.min, children: [
+            Flexible(
+              flex: 5,
+              child: Padding(
+                padding: EdgeInsets.only(left: _screenSize.width * 0.1),
+                child: Text(msg),
+              ),
+            ),
+            Flexible(
+              child: TextButton(
+                onPressed: () => AppRouterDelegate.nextRoute('/privacy'),
+                child: Text(
+                  _smallScreen ? 'Info' : 'Learn More',
+                  style: const TextStyle(
+                      color: Colors.blue, decoration: TextDecoration.underline),
+                ),
+              ),
+            ),
+            Flexible(
+              flex: 2,
+              child: Padding(
+                padding: EdgeInsets.only(left: _screenSize.width * 0.1),
+                child: Align(
+                  alignment: Alignment.centerRight,
+                  child: OutlinedButton(
+                    onPressed: () {
+                      con.setOverlay(shown: false);
+                      setState(() {});
+                    },
+                    child: const Text(
+                      'OK',
+                      style: TextStyle(
+                        color: Colors.black,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ]),
+        ),
+      ),
+    );
+    return StackWidgetProperties(
+      alignment: Alignment.bottomRight,
+      child: widget,
+    );
   }
 
   @override
@@ -59,6 +152,10 @@ class HomePageSmallController<T> extends WebPageController {
   @override
   void initWidget() {
     //
+
+    // Determine if an overlay is to be displayed.
+    showOverlay = Prefs.getBool('showOverlay', true);
+
     projects = HowProjectsWork();
     useCase = UseCaseExample(bottomBar: false);
     paradox = ProgrammingParadox();
@@ -72,12 +169,16 @@ class HomePageSmallController<T> extends WebPageController {
       }
       _lastOffset = offset;
     });
+    flutterUIs = FlutterUIs();
   }
 
   late HowProjectsWork projects;
   late UseCaseExample useCase;
   late ProgrammingParadox paradox;
   late InitialDisclosure disclose;
+  late FlutterUIs flutterUIs;
+  late bool showOverlay;
+
   double _lastOffset = 0;
 
   List<Widget> children04(BuildContext context) {
@@ -148,6 +249,12 @@ class HomePageSmallController<T> extends WebPageController {
     );
     children.add(DartPackages());
     children.addAll(disclose.children05(context));
+    children.add(flutterUIs.coverPage(context));
     return children;
+  }
+
+  Future<bool> setOverlay({bool shown = true}) {
+    showOverlay = shown;
+    return Prefs.setBool('showOverlay', shown);
   }
 }